Scale mutation rate with fitness
Add calculation at start of every main while loop iteration that scales the mutation rate down as fitness moves toward 0. Attempting to keep generations from getting stuck at low fitness (1 or 2) and then taking many more generations to converge on the target string. Not sure if this is the best solution, but trying it.